+++ packages/data_layer/test/sources/source_list_watch_test.dart
@@ -96,8 +96,13 @@ class CountingWatchableSource extends Source<TestModel>
   }
 
   @override
-  Future<DeleteResult<TestModel>> delete(
+  Future<DeleteResult<TestModel>> deleteItem(
     DeleteOperation<TestModel> operation,
+  ) async => throw UnimplementedError();
+
+  @override
+  Future<DeleteResult<TestModel>> deleteItems(
+    DeleteListOperation<TestModel> operation,
   ) async => throw UnimplementedError();
 
   @override